+++ src/lib/state/groups.ts
@@ -75,14 +75,14 @@ class GroupStore {
     if (t === GroupType.LFG) {
       if (!this._groups.has(g.id)) {
         this._groups.set(g.id, g as Group)
-        schedule.scheduleJob(`remove_group:${g.id}`, addHour(new Date(), 12), () =>
+        schedule.scheduleJob(`remove_group:${g.id}`, addHour(new Date(), 8), () =>
           this.remove(g.id, GroupType.LFG)
         )
       }
     } else {
       if (!this._flights.has(g.id)) {
         this._flights.set(g.id, g as Flight)
-        schedule.scheduleJob(`remove_group:${g.id}`, addHour(new Date(), 12), () =>
+        schedule.scheduleJob(`remove_group:${g.id}`, addHour(new Date(), 8), () =>
           this.remove(g.id, GroupType.Flight)
         )
       }