+++ .github/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - ADD `pusher-js` capabilities for hooking in the event stream served by the UO Authenticator for providing and revoking user permissions
 - UPDATE calendar event notification intervals
+- UPDATE temporarily disable `!ready` command until primary server features are updated to be compatible
 - REMOVE FTP related environment variable dependencies since transition to forums API for mission searching
 
 ## v2.8.0