+++ .github/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v2.9.0
+
+- ADD `pusher-js` capabilities for hooking in the event stream served by the UO Authenticator for providing and revoking user permissions
+- ADD `Steel Beasts` Discord role to the approved list for `join_group` and `leave_group`. **Replace all spaces with underscores in role names when running the command**
+- UPDATE flight and LFG groups expire after 8 hours
+- UPDATE calendar event notification intervals
+- UPDATE temporarily disable `!ready` command until primary server features are updated to be compatible
+- REMOVE FTP related environment variable dependencies since transition to forums API for mission searching
+
 ## v2.8.0
 
 - ADD separate store to house and manage Discord channel functionality and IDs